Split elster data representation into section-specific files.
